+++ tests/testthat/test_ObjectiveFSelect.R
@@ -12,7 +12,7 @@ test_that("ObjectiveFSelect", {
     list("x1" = FALSE, "x2" = TRUE, "x3" = TRUE, "x4" = TRUE))
 
   z = obj$eval_many(xss)
-  expect_data_table(z, nrows = 2, ncols = 2)
+  expect_data_table(z, nrows = 2, ncols = 3)
   expect_equal(obj$archive$benchmark_result$resample_result(1)$learners[[1]]$model$select$selection, c("x1", "x3", "x4"))
   expect_equal(obj$archive$benchmark_result$resample_result(2)$learners[[1]]$model$select$selection, c("x2", "x3", "x4"))
 })
@@ -31,7 +31,7 @@ test_that("ObjectiveFSelect works with multiple measures", {
     list("x1" = FALSE, "x2" = TRUE, "x3" = TRUE, "x4" = TRUE))
 
   z = obj$eval_many(xss)
-  expect_data_table(z, nrows = 2, ncols = 3)
+  expect_data_table(z, nrows = 2, ncols = 4)
 })
 
 test_that("ObjectiveFSelect works with store_models", {