+++ NEWS.md
@@ -1,6 +1,8 @@
 # mlr3fselect 0.5.1.9000
 
 - Adds `fselect()`, `auto_fselector()` and `fselect_nested()` sugar functions.
+- Adds `extract_inner_fselect_results()` and `extract_inner_fselect_archives()`  
+  helper function to extract inner feature selection results and archives.
 
 # mlr3fselect 0.5.1